+++ custom_components/c4_audio/media_player.py
@@ -51,9 +51,12 @@ class C4ZoneMediaPlayer(CoordinatorEntity[C4AudioCoordinator], MediaPlayerEntity
         if switch is not None:
             self.async_on_remove(switch.async_add_listener(self._handle_switch_update))
         registry = er.async_get(self.hass)
-        registry.async_update_entity(
-            self.entity_id, area_id=self.coordinator.zone_area_id(self._zone)
-        )
+        if self.coordinator.is_matrix:
+            registry.async_update_entity(self.entity_id, area_id=None)
+        else:
+            area_id = self.coordinator.zone_area_id(self._zone)
+            if area_id:
+                registry.async_update_entity(self.entity_id, area_id=area_id)
 
     @callback
     def _handle_switch_update(self) -> None: